// src/api/model/accountDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AccountRoleTypeApi } from './accountRoleType';


/**
 * Account DTO type for response
 */
export interface AccountDtoApi { 
    /**
     * Username for login
     */
    username?: string;
    /**
     * Account role type
     */
    roleType?: AccountRoleTypeApi;
}
export namespace AccountDtoApi {
}