// src/api/model/accountSignInDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Account DTO type for sign in.
 */
export interface AccountSignInDtoApi { 
    /**
     * Username for login
     */
    username?: string;
    /**
     * Password
     */
    password?: string;
}