// src/api/model/accountSignUpDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Account DTO type for sign up.
 */
export interface AccountSignUpDtoApi { 
    /**
     * Username for login
     */
    username?: string;
    /**
     * Password
     */
    password?: string;
}