// src/api/model/uriDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Common URI data
 */
export interface UriDtoApi { 
    /**
     * URI value
     */
    uri?: string;
    /**
     * URI title
     */
    title?: string;
    /**
     * URI description
     */
    description?: string;
}